Tottally rework ALL 3 LAYERS. Autocreate table in many-to-many relations, removed useless entities

+++ EdProject.PresentationLayer/Middleware/ExceptionMiddleware.cs
@@ -25,7 +25,7 @@ namespace EdProject.PresentationLayer.Middleware
             }
             catch (CustomException ex)
             {
-                httpContext.Response.StatusCode = ex.StatusCode;
+                httpContext.Response.StatusCode = (int)ex.StatusCode;
                 await httpContext.Response.WriteAsync(ex.Message);
             }
             catch (Exception ex)
